requirements.txt
httpx>=0.27
websockets>=13.0
python-dotenv>=1.0
